fix error for heredoc

+++ main.c
@@ -3,14 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mehdimirzaie <mehdimirzaie@student.42.f    +#+  +:+       +#+        */
+/*   By: mmirzaie <mmirzaie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/29 10:11:10 by mmirzaie          #+#    #+#             */
-/*   Updated: 2023/09/07 19:10:08 by mehdimirzai      ###   ########.fr       */
+/*   Updated: 2023/09/08 12:17:02 by mmirzaie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "pipex.h"
+#include "next_line.h"
 
 void	middle_child(const t_args args, int pipes[][2], int i)
 {
@@ -25,10 +26,9 @@ void	middle_child(const t_args args, int pipes[][2], int i)
 			close(pipes[j][1]);
 		j++;
 	}
-	dup2(pipes[i][0], STDIN_FILENO);
-	dup2(pipes[i + 1][1], STDOUT_FILENO);
-	close(pipes[i][0]);
-	close(pipes[i + 1][1]);
+	if ((redirect_input(pipes[i][0], STDIN_FILENO) == -1)
+		|| redirect_output(pipes[i + 1][1], STDOUT_FILENO) == -1)
+		exit(EXIT_FAILURE);
 	pipex(args.av[i + 3], args.env);
 }
 
@@ -40,12 +40,11 @@ void	child(const t_args args, int pipes[][2])
 	file = open(args.av[1], O_RDONLY);
 	if (file == -1)
 		exit(127);
-	dup2(file, STDIN_FILENO);
-	dup2(pipes[0][1], STDOUT_FILENO);
-	close(pipes[0][1]);
-	close(file);
+	if ((redirect_input(file, STDIN_FILENO) == -1)
+		|| redirect_output(pipes[0][1], STDOUT_FILENO) == -1)
+		exit(EXIT_FAILURE);
 	if (args.temp_file_created == true)
-		if (unlink(args.av[1]) == -1) // Remove the temporary file
+		if (unlink(args.av[1]) == -1)
 			perror("unlink");
 	pipex(args.av[2], args.env);
 }
@@ -62,16 +61,13 @@ void	parent(const t_args args, int pipes[][2])
 		file = open(args.av[args.ac - 1], O_WRONLY | O_CREAT | O_TRUNC, 0664);
 	if (file == -1)
 		exit(127);
-	dup2(file, STDOUT_FILENO);
-	dup2(pipes[args.process_num][0], STDIN_FILENO);
-	close(file);
-	close(pipes[args.process_num][0]);
-	close(pipes[0][1]);
-	close(pipes[args.process_num][0]);
-	pipex(args.av[args.ac - 2], args.env);
+	if ((redirect_input(file, STDOUT_FILENO) == -1)
+		|| redirect_output(pipes[args.process_num][0], STDIN_FILENO) == -1)
+		exit(EXIT_FAILURE);
 	i = 0;
 	while (i++ < args.process_num)
 		wait(NULL);
+	pipex(args.av[args.ac - 2], args.env);
 }
 
 void	master(const t_args args, int pipes[][2])
@@ -92,10 +88,7 @@ void	master(const t_args args, int pipes[][2])
 	parent(args, pipes);
 }
 
-#define MAX_PROCESS_NUM 100
-#include "next_line.h"
-
-int open_pipes(t_args args, int pipes[][2])
+int	open_pipes(t_args args, int pipes[][2])
 {
 	int	i;
 
@@ -131,74 +124,76 @@ int	create_middle_process(t_args args, int pipes[][2])
 	return (0);
 }
 
+void	new_av(t_args args, char *filename)
+{
+	int	i;
+
+	i = 2;
+	args.av[1] = filename;
+	printf("\n%s\n", args.av[1]);
+	while (i < (args.ac - 1))
+	{
+		args.av[i] = args.av[i + 1];
+		printf("%s\n", args.av[i]);
+		i++;
+	}
+	printf("\n");
+	args.ac = i;
+	printf("%d\n", args.ac);
+	args.process_num = args.ac - 5;
+	printf("%d\n", args.process_num);
+}
+
+int	heredoc(t_args args, int pipes[][2])
+{
+	const char	*template = "/tmp/mytempfileXXXXXX";
+	int			fd;
+	char		*filename;
+	char		*lines;
+	int			limiter_len;
+
+	filename = ft_strdup(template);
+	fd = open(filename, O_CREAT | O_RDWR | O_EXCL, S_IRUSR | S_IWUSR);
+	if (fd == -1)
+		exit(127);
+	args.temp_file_created = true;
+	limiter_len = ft_strlen(args.av[2]);
+	lines = get_next_line(STDIN_FILENO);
+	while (ft_strncmp(lines, args.av[2], limiter_len) != 0)
+	{
+		ft_putstr_fd(lines, fd);
+		free(lines);
+		lines = get_next_line(STDIN_FILENO);
+	}
+	new_av(args, filename);
+	// args.av[1] = filename;
+	// args.av[2] = args.av[3];
+	// args.av[3] = args.av[4];
+	// args.av[4] = args.av[5];
+	// args.ac = 5;
+	// args.process_num = args.ac - 5;
+	if (open_pipes(args, pipes) == 1 || create_middle_process(args, pipes) == 2)
+		return (1);
+	master(args, pipes);
+	return (0);
+}
+
 int	main(int ac, char **av, char **env)
 {
-	int		process_num = ac - 5;
 	int		pids[MAX_PROCESS_NUM];
 	int		pipes[MAX_PROCESS_NUM][2];
-	t_args	args = {av, ac, process_num, env, pids, false};
+	t_args	args;
 
+	args = (t_args){.av = av, .ac = ac,
+		.process_num = ac - 5,
+		.env = env, .pids = pids, .temp_file_created = false};
 	if (*env == NULL)
 		return (0);
 	if (ac < 3)
 		return (127);
-	
 	if (ft_strncmp(av[1], "here_doc", 8) == 0)
-	{
-		char* template = "/tmp/mytempfileXXXXXX";
-		int fd;
-		char *filename = strdup(template);
-
-		fd = open(filename, O_CREAT | O_RDWR | O_EXCL, S_IRUSR | S_IWUSR);
-		if (fd == -1)
-			exit(127);
-		args.temp_file_created = true;
-		int	limiter_len = ft_strlen(av[2]);
-		char *lines;
-		lines = get_next_line(STDIN_FILENO);
-		while (ft_strncmp(lines, av[2], limiter_len) != 0)
-		{
-			ft_putstr_fd(lines, fd);
-			free(lines);
-			lines = get_next_line(STDIN_FILENO);
-		}
-		av[1] = filename;
-		av[2] = av[3];
-		av[3] = av[4];
-		av[4] = av[5];
-		args.ac = 5;
-		args.process_num = ac - 5;
-		if (open_pipes(args, pipes) == 1)
-			return (1);
-		if (create_middle_process(args, pipes) == 2)
-			return (2);
-		master(args, pipes);
-	}
-	if (ac == 4)
-	{
-		int fd[2];
-		pipe(fd);
-		if (fork() == 0){
-			close(fd[0]);
-			int file = open(av[1], O_RDONLY);
-			if (file == -1)
-				exit(127);
-			dup2(file, STDIN_FILENO);
-			dup2(fd[1], STDOUT_FILENO);
-			close(fd[1]);
-			close(file);
-			pipex(av[2], env);
-		}
-		close(fd[1]);
-		dup2(fd[0], STDIN_FILENO);
-		int file = open(av[3], O_WRONLY | O_CREAT | O_TRUNC, 0664);
-		if (file == -1)
-			exit(127);
-		dup2(file, STDOUT_FILENO);
-		close(file);
-		close(fd[0]);
-		pipex(av[3], env);
-	}
+		if (heredoc(args, pipes) != 0)
+			exit(EXIT_FAILURE);
 	if (open_pipes(args, pipes) == 1)
 		return (1);
 	if (create_middle_process(args, pipes) == 2)
